refactor(frontend): support sssr, group routes
