+++ frontend/next.config.mjs
@@ -1,9 +1,5 @@
-// NextJS dev server will not support direct URLs otherwise
-const shouldBuildSPA = process.env.NODE_ENV === 'production';
-
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    output: shouldBuildSPA ? 'export' : undefined,
     distDir: './build',
     images: {
         remotePatterns: [
