+++ frontend/src/app/(blog)/blog/common/Footer.tsx
