+++ frontend/src/app/(blog)/blog/dojo-digest/vol-8/clock-graph.png
