+++ frontend/src/app/(blog)/blog/page.tsx
