+++ frontend/src/app/(blog)/blog/tactics-test/sample-problem.png
