+++ frontend/src/app/(legacy)/[[...slug]]/client.tsx
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic';
 import { useReportWebVitals } from 'next/web-vitals';
 import ReactGA from 'react-ga4';
 
-const App = dynamic(() => import('../../App'), { ssr: false });
+const App = dynamic(() => import('@/App'), { ssr: false });
 
 ReactGA.initialize('G-9VPNTDELD2');
