+++ frontend/src/app/layout.tsx
@@ -1,16 +1,4 @@
-import { ApiProvider } from '@/api/Api';
-import { AuthProvider } from '@/auth/Auth';
-import Navbar from '@/navbar/Navbar';
-import ThemeProvider from '@/style/ThemeProvider';
-import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
-import type { Metadata } from 'next';
 import './globals.css';
-
-export const metadata: Metadata = {
-    title: 'ChessDojo',
-    description:
-        'View the scoreboard, schedule meetings and more for the ChessDojo Training Program',
-};
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
     return (
@@ -20,19 +8,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
                 <link rel='manifest' href='/manifest.json' />
             </head>
             <body>
-                <div id='root'>
-                    <AppRouterCacheProvider>
-                        <ThemeProvider>
-                            <AuthProvider>
-                                <ApiProvider>
-                                    <Navbar />
-
-                                    {children}
-                                </ApiProvider>
-                            </AuthProvider>
-                        </ThemeProvider>
-                    </AppRouterCacheProvider>
-                </div>
+                <div id='root'>{children}</div>
             </body>
         </html>
     );
